+++ cosi/lab1/mod/lab2/exponentialUi.ts
@@ -49,7 +49,7 @@ let defaultData: IFormData = {
 
         (<webix.ui.button>$$(exponentialRunButtonId)).attachEvent("onItemClick", function () {
             let data = form.getValues() as IFormDataGet,
-                results = utils.exponentialDistribution(parseInt(data.alpha), parseInt(data.count)),
+                results = utils.exponentialDistribution(parseFloat(data.alpha), parseInt(data.count)),
                 mX = utils.getMx(results),
                 dX = utils.getDx(results, mX),
                 chartData = utils.getChartData(results);
